+++ packages/datadog_flutter_plugin/android/src/main/kotlin/com/datadoghq/flutter/DatadogRumPlugin.kt
@@ -6,10 +6,7 @@
 package com.datadoghq.flutter
 
 import android.content.Context
-import android.os.Handler
-import android.os.Looper
 import android.util.Log
-import com.datadog.android.Datadog
 import com.datadog.android.api.SdkCore
 import com.datadog.android.core.configuration.BatchSize
 import com.datadog.android.core.configuration.UploadFrequency
@@ -25,12 +22,6 @@ import com.datadog.android.rum.RumPerformanceMetric
 import com.datadog.android.rum.RumResourceKind
 import com.datadog.android.rum._RumInternalProxy
 import com.datadog.android.rum.configuration.VitalsUpdateFrequency
-import com.datadog.android.rum.event.ViewEventMapper
-import com.datadog.android.rum.model.ActionEvent
-import com.datadog.android.rum.model.ErrorEvent
-import com.datadog.android.rum.model.LongTaskEvent
-import com.datadog.android.rum.model.ResourceEvent
-import com.datadog.android.rum.model.ViewEvent
 import com.datadog.android.rum.tracking.ViewTrackingStrategy
 import com.datadog.android.telemetry.model.TelemetryConfigurationEvent
 import io.flutter.embedding.engine.plugins.FlutterPlugin
@@ -38,14 +29,9 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.Result
 import java.lang.ClassCastException
-import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
-import kotlin.system.measureNanoTime
 
-@Suppress("StringLiteralDuplication")
-class DatadogRumPlugin internal constructor(
-    rumInstance: RumMonitor? = null
-) : MethodChannel.MethodCallHandler {
+class DatadogRumPlugin : MethodChannel.MethodCallHandler {
     companion object RumParameterNames {
         const val PARAM_AT = "at"
         const val PARAM_DURATION = "duration"
@@ -69,9 +55,8 @@ class DatadogRumPlugin internal constructor(
         // See DatadogSdkPlugin's description of this same member
         private var previousConfiguration: Map<String, Any?>? = null
 
-        val instance: DatadogRumPlugin by lazy {
-            DatadogRumPlugin()
-        }
+        // Static instance of the event mapper
+        internal val eventMapper: DatadogRumEventMapper = DatadogRumEventMapper()
 
         // For testing purposes only
         internal fun resetConfig() {
@@ -82,12 +67,8 @@ class DatadogRumPlugin internal constructor(
     private lateinit var channel: MethodChannel
     private lateinit var binding: FlutterPlugin.FlutterPluginBinding
 
-    var rum: RumMonitor? = rumInstance
+    var rum: RumMonitor? = null
         private set
-
-    val mapperPerf = PerformanceTracker()
-    val mapperPerfMainThread = PerformanceTracker()
-    var mapperTimeouts = 0
 
     // Might need a better way to deal with this. There's weird shared responsibility for
     // telemetry between the core and RUM.
@@ -96,11 +77,17 @@ class DatadogRumPlugin internal constructor(
     fun attachToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
         channel = MethodChannel(flutterPluginBinding.binaryMessenger, "datadog_sdk_flutter.rum")
         channel.setMethodCallHandler(this)
+        eventMapper.addChannel(channel)
 
         binding = flutterPluginBinding
+
+        if (GlobalRumMonitor.isRegistered()) {
+            rum = GlobalRumMonitor.get()
+        }
     }
 
     fun detachFromEngine() {
+        eventMapper.removeChannel(channel)
         channel.setMethodCallHandler(null)
     }
 
@@ -228,57 +215,7 @@ class DatadogRumPlugin internal constructor(
         config: Map<String, Any?>,
         configBuilder: RumConfiguration.Builder
     ): RumConfiguration.Builder {
-        fun optionIsSet(key: String): Boolean {
-            return config[key] as? Boolean ?: false
-        }
-
-        if (optionIsSet("attachViewEventMapper")) {
-            configBuilder.setViewEventMapper(
-                object : ViewEventMapper {
-                    override fun map(event: ViewEvent): ViewEvent {
-                        return mapViewEvent(event)
-                    }
-                }
-            )
-        }
-        if (optionIsSet("attachActionEventMapper")) {
-            configBuilder.setActionEventMapper(
-                object : EventMapper<ActionEvent> {
-                    override fun map(event: ActionEvent): ActionEvent? {
-                        return mapActionEvent(event)
-                    }
-                }
-            )
-        }
-        if (optionIsSet("attachResourceEventMapper")) {
-            configBuilder.setResourceEventMapper(
-                object : EventMapper<ResourceEvent> {
-                    override fun map(event: ResourceEvent): ResourceEvent? {
-                        return mapResourceEvent(event)
-                    }
-                }
-            )
-        }
-        if (optionIsSet("attachErrorEventMapper")) {
-            configBuilder.setErrorEventMapper(
-                object : EventMapper<ErrorEvent> {
-                    override fun map(event: ErrorEvent): ErrorEvent? {
-                        return mapErrorEvent(event)
-                    }
-                }
-            )
-        }
-        if (optionIsSet("attachLongTaskEventMapper")) {
-            configBuilder.setLongTaskEventMapper(
-                object : EventMapper<LongTaskEvent> {
-                    override fun map(event: LongTaskEvent): LongTaskEvent? {
-                        return mapLongTaskEvent(event)
-                    }
-                }
-            )
-        }
-
-        return configBuilder
+        return eventMapper.attachMappers(config, configBuilder)
     }
 
     private fun startView(call: MethodCall, result: Result) {
@@ -504,266 +441,6 @@ class DatadogRumPlugin internal constructor(
     private fun stopSession(call: MethodCall, result: Result) {
         rum?.stopSession()
         result.success(null)
-    }
-
-    @Suppress("TooGenericExceptionCaught")
-    internal fun <T> callEventMapper(
-        mapperName: String,
-        event: T,
-        encodedEvent: Map<String, Any?>,
-        completion: (Map<String, Any?>?, T) -> T?
-    ): T? {
-        var modifiedJson: Map<String, Any?>? = encodedEvent
-        val latch = CountDownLatch(1)
-
-        val handler = Handler(Looper.getMainLooper())
-        handler.post {
-            val perf = measureNanoTime {
-                try {
-                    channel.invokeMethod(
-                        mapperName,
-                        mapOf(
-                            "event" to encodedEvent
-                        ),
-                        object : Result {
-                            @Suppress("UNCHECKED_CAST")
-                            override fun success(result: Any?) {
-                                modifiedJson = (result as? Map<String, Any?>)
-                                latch.countDown()
-                            }
-
-                            override fun error(
-                                errorCode: String,
-                                errorMessage: String?,
-                                errorDetails: Any?
-                            ) {
-                                latch.countDown()
-                            }
-
-                            override fun notImplemented() {
-                                Datadog._internalProxy()._telemetry.error(
-                                    "$mapperName returned notImplemented."
-                                )
-                                latch.countDown()
-                            }
-                        }
-                    )
-                } catch (e: Exception) {
-                    Datadog._internalProxy()._telemetry.error(
-                        "Attempting call $mapperName failed.",
-                        e
-                    )
-                    latch.countDown()
-                }
-            }
-            mapperPerfMainThread.addSample(perf)
-        }
-
-        try {
-            // Stalls until the method channel finishes
-            if (!latch.await(1, TimeUnit.SECONDS)) {
-                Datadog._internalProxy()._telemetry.debug("$mapperName timed out")
-                return event
-            }
-
-            if (modifiedJson?.containsKey("_dd.mapper_error") == true) {
-                return event
-            }
-
-            return completion(modifiedJson, event)
-        } catch (e: InterruptedException) {
-            Datadog._internalProxy()._telemetry.debug(
-                "Latch await was interrupted. Returning unmodified event."
-            )
-        } catch (e: Exception) {
-            Datadog._internalProxy()._telemetry.error(
-                "Unknown exception attempting to deserialize mapped log event." +
-                    " Returning unmodified event.",
-                e
-            )
-        }
-
-        return event
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    internal fun mapViewEvent(event: ViewEvent): ViewEvent {
-        var result: ViewEvent
-        val perf = measureNanoTime {
-            var jsonEvent = event.toJson().asFlutterMap()
-            jsonEvent = normalizeExtraUserInfo(jsonEvent)
-
-            result = callEventMapper("mapViewEvent", event, jsonEvent) { encodedResult, event ->
-                (encodedResult?.get("view") as? Map<String, Any?>)?.let {
-                    event.view.name = it["name"] as? String
-                    event.view.referrer = it["referrer"] as? String
-                    event.view.url = it["url"] as String
-                }
-
-                event
-            } ?: event
-        }
-        mapperPerf.addSample(perf)
-
-        return result
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    internal fun mapActionEvent(event: ActionEvent): ActionEvent? {
-        val result: ActionEvent?
-        val perf = measureNanoTime {
-            var jsonEvent = event.toJson().asFlutterMap()
-            jsonEvent = normalizeExtraUserInfo(jsonEvent)
-
-            result = callEventMapper("mapActionEvent", event, jsonEvent) { encodedResult, event ->
-                if (encodedResult == null) {
-                    null
-                } else {
-                    (encodedResult["action"] as? Map<String, Any?>)?.let {
-                        val encodedTarget = it["target"] as? Map<String, Any?>
-                        if (encodedTarget != null) {
-                            event.action.target?.name = encodedTarget["name"] as String
-                        }
-                    }
-
-                    (encodedResult["view"] as? Map<String, Any?>)?.let {
-                        event.view.name = it["name"] as? String
-                        event.view.referrer = it["referrer"] as? String
-                        event.view.url = it["url"] as String
-                    }
-
-                    event
-                }
-            }
-        }
-        mapperPerf.addSample(perf)
-
-        return result
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    internal fun mapResourceEvent(event: ResourceEvent): ResourceEvent? {
-        var result: ResourceEvent?
-        val perf = measureNanoTime {
-            var jsonEvent = event.toJson().asFlutterMap()
-            jsonEvent = normalizeExtraUserInfo(jsonEvent)
-
-            result = callEventMapper("mapResourceEvent", event, jsonEvent) { encodedResult, event ->
-                if (encodedResult == null) {
-                    null
-                } else {
-                    (encodedResult["resource"] as? Map<String, Any?>)?.let {
-                        event.resource.url = it["url"] as String
-                    }
-
-                    (encodedResult["view"] as? Map<String, Any?>)?.let {
-                        event.view.name = it["name"] as? String
-                        event.view.referrer = it["referrer"] as? String
-                        event.view.url = it["url"] as String
-                    }
-
-                    event
-                }
-            }
-        }
-        mapperPerf.addSample(perf)
-
-        return result
-    }
-
-    @Suppress("ComplexMethod", "UNCHECKED_CAST")
-    internal fun mapErrorEvent(event: ErrorEvent): ErrorEvent? {
-        var result: ErrorEvent?
-        val perf = measureNanoTime {
-            var jsonEvent = event.toJson().asFlutterMap()
-            jsonEvent = normalizeExtraUserInfo(jsonEvent)
-
-            result = callEventMapper("mapErrorEvent", event, jsonEvent) { encodedResult, event ->
-                if (encodedResult == null) {
-                    null
-                } else {
-                    (encodedResult["error"] as? Map<String, Any?>)?.let { encodedError ->
-                        val encodedCauses = encodedError["causes"] as? List<Map<String, Any?>>
-                        if (encodedCauses != null) {
-                            event.error.causes?.let { causes ->
-                                if (causes.count() == encodedCauses.count()) {
-                                    causes.forEachIndexed { i, cause ->
-                                        cause.message = encodedCauses[i]["message"] as? String ?: ""
-                                        cause.stack = encodedCauses[i]["stack"] as? String
-                                    }
-                                }
-                            }
-                        } else {
-                            event.error.causes = null
-                        }
-
-                        val encodedResource = encodedError["resource"] as? Map<String, Any?>
-                        if (encodedResource != null) {
-                            event.error.resource?.url = encodedResource["url"] as? String ?: ""
-                        }
-
-                        event.error.stack = encodedError["stack"] as? String
-                        event.error.fingerprint = encodedError["fingerprint"] as? String
-                    }
-
-                    (encodedResult["view"] as? Map<String, Any?>)?.let {
-                        event.view.name = it["name"] as? String
-                        event.view.referrer = it["referrer"] as? String
-                        event.view.url = it["url"] as String
-                    }
-
-                    event
-                }
-            }
-        }
-        mapperPerf.addSample(perf)
-
-        return result
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    internal fun mapLongTaskEvent(event: LongTaskEvent): LongTaskEvent? {
-        var result: LongTaskEvent?
-        val perf = measureNanoTime {
-            var jsonEvent = event.toJson().asFlutterMap()
-            jsonEvent = normalizeExtraUserInfo(jsonEvent)
-
-            result = callEventMapper("mapLongTaskEvent", event, jsonEvent) { encodedResult, event ->
-                if (encodedResult == null) {
-                    null
-                } else {
-                    (encodedResult["view"] as? Map<String, Any?>)?.let {
-                        event.view.name = it["name"] as? String
-                        event.view.referrer = it["referrer"] as? String
-                        event.view.url = it["url"] as String
-                    }
-
-                    event
-                }
-            }
-        }
-        mapperPerf.addSample(perf)
-
-        return result
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    private fun normalizeExtraUserInfo(encodedEvent: Map<String, Any?>): Map<String, Any?> {
-        val reservedKeys = setOf("email", "id", "name")
-        // Pull out user information
-        val mutableEvent = encodedEvent.toMutableMap()
-        (mutableEvent["usr"] as? Map<String, Any?>)?.let { usr ->
-            val mutableUsr = usr.toMutableMap()
-            val extraUserInfo = mutableMapOf<String, Any?>()
-            usr.filter { !reservedKeys.contains(it.key) }.forEach {
-                extraUserInfo[it.key] = it.value
-                mutableUsr.remove(it.key)
-            }
-            mutableUsr["usr_info"] = extraUserInfo
-            mutableEvent["usr"] = mutableUsr
-        }
-
-        return mutableEvent
     }
 }
